Add files via upload
Smart-Car-Parking-System
Smart Car Parking System using Arduino Uno

ABSTRACT

Smart car parking project aims at providing a confusion free and easy parking. This project helps the drivers of the cars to park their vehicles with minimum wastage of time with accurate information of the availability of the space to park It includes an Arduino Uno as the microcontroller unit to which the servo motors, LCD display ultrasonic sensors (HC-05) are interfaced. The LCD displays the availability of the space, the ultrasonic sensors keeps the check of the number of cars entering and exiting the parking space. The ultrasonic sensors detect the availability of the parking space.

THEORY

An automated car parking system is a process through which car parking can be done more efficiently and easily than manual method. The system will provide the user better services. The system counts the number of cars in the garage and checks if there’s any vacancy. There’s an entry and exit path. When vehicle enters, the display shows the number of cars inside. When any vehicle leaves, the count decreases and shown on display. If the garage is full. The display will show a message regarding that. This whole process includes the use of Arduino, Display and sonar. The sonar detects whether the vehicle is entering or leaving. The report then showed on display.

Hardware Design

Arduino UNO
Ultrasonic Sensor
LCD Screen
Bread Board
Power Supply
Connecting Wires
Block Diagram